+++ src/main/java/me/figo/FigoSession.java
@@ -23,23 +23,9 @@
 package me.figo;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.lang.reflect.Type;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.net.URLEncoder;
 import java.util.List;
-import java.util.Scanner;
-import java.nio.charset.Charset;
-
-import javax.net.ssl.HttpsURLConnection;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-
-import me.figo.FigoException.ErrorResponse;
 import me.figo.internal.AccountOrderRequest;
-import me.figo.internal.FigoTrustManager;
-import me.figo.internal.GsonAdapter;
 import me.figo.internal.SetupAccountRequest;
 import me.figo.internal.SubmitPaymentRequest;
 import me.figo.internal.SyncTokenRequest;
@@ -63,9 +49,6 @@ import me.figo.models.Payment;
 import me.figo.models.Transaction;
 import me.figo.models.User;
 
-import com.google.gson.Gson;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
 import java.util.Collections;
 
 /**
@@ -73,7 +56,7 @@ import java.util.Collections;
  * 
  * @author Stefan Richter
  */
-public class FigoSession {
+public class FigoSession extends FigoApi {
     
     public enum PendingTransactions {
         INCLUDED,
@@ -84,14 +67,6 @@ public class FigoSession {
         VISITED,
         NOT_VISITED
     }
-
-    private final String apiEndpoint;
-
-    private final String access_token;
-
-    private final int timeout;
-    
-    private final Charset UTF_8 = Charset.forName("UTF-8");
 
     /**
      * Creates a FigoSession instance
@@ -126,109 +101,107 @@ public class FigoSession {
      *            which endpoint to use (customize for different figo deployment)
      */
     public FigoSession(String accessToken, int timeout, String apiEndpoint) {
-        this.access_token = accessToken;
-        this.timeout = timeout;
-        this.apiEndpoint = apiEndpoint;
+        super(apiEndpoint, "Bearer " + accessToken, timeout);
     }
 
-    /**
-     * Query the figo API via HTTP
-     * 
-     * @param path
-     *            Endpoint to query
-     * @param data
-     *            Payload to send
-     * @param method
-     *            HTTP verb to use
-     * @param typeOfT
-     *            Type of expected response
-     * @param <T>
-     *            Type of expected response
-     * @return decoded response
-     */
-    protected <T> T queryApi(String path, Object data, String method, Type typeOfT) throws FigoException, IOException {
-        // configure URL connection, i.e. the HTTP request
-        HttpURLConnection connection = (HttpURLConnection) (new URL(apiEndpoint + path)).openConnection();
-        connection.setConnectTimeout(timeout);
-        connection.setReadTimeout(timeout);
+//    /**
+//     * Query the figo API via HTTP
+//     * 
+//     * @param path
+//     *            Endpoint to query
+//     * @param data
+//     *            Payload to send
+//     * @param method
+//     *            HTTP verb to use
+//     * @param typeOfT
+//     *            Type of expected response
+//     * @param <T>
+//     *            Type of expected response
+//     * @return decoded response
+//     */
+//    protected <T> T queryApi(String path, Object data, String method, Type typeOfT) throws FigoException, IOException {
+//        // configure URL connection, i.e. the HTTP request
+//        HttpURLConnection connection = (HttpURLConnection) (new URL(apiEndpoint + path)).openConnection();
+//        connection.setConnectTimeout(timeout);
+//        connection.setReadTimeout(timeout);
+//
+//        if (connection instanceof HttpsURLConnection) {
+//            // Setup and install the trust manager
+//            try {
+//                final SSLContext sc = SSLContext.getInstance("SSL");
+//                sc.init(null, new TrustManager[] { new FigoTrustManager() }, new java.security.SecureRandom());
+//                ((HttpsURLConnection) connection).setSSLSocketFactory(sc.getSocketFactory());
+//            } catch (NoSuchAlgorithmException e) {
+//                throw new IOException("Connection setup failed", e);
+//            } catch (KeyManagementException e) {
+//                throw new IOException("Connection setup failed", e);
+//            }
+//        }
+//
+//        connection.setRequestMethod(method);
+//        connection.setRequestProperty("Authorization", "Bearer " + this.access_token);
+//        connection.setRequestProperty("Accept", "application/json");
+//        connection.setRequestProperty("Content-Type", "application/json");
+//
+//        // add payload
+//        if (data != null) {
+//            String encodedData = createGson().toJson(data);
+//
+//            connection.setDoOutput(true);
+//            connection.getOutputStream().write(encodedData.getBytes(UTF_8));
+//        }
+//
+//        // process response
+//        int code = connection.getResponseCode();
+//        if (code >= 200 && code < 300) {
+//            if (typeOfT == null)
+//                return null;
+//            else
+//                return handleResponse(connection.getInputStream(), typeOfT);
+//        } else if (code == 400) {
+//            throw new FigoException((ErrorResponse) handleResponse(connection.getErrorStream(), FigoException.ErrorResponse.class));
+//        } else if (code == 401) {
+//            throw new FigoException("access_denied", "Access Denied");
+//        } else if (code == 404) {
+//            return null;
+//        } else {
+//            // return decode(connection.getErrorStream(), resultType);
+//            throw new FigoException("internal_server_error", "We are very sorry, but something went wrong");
+//        }
+//    }
 
-        if (connection instanceof HttpsURLConnection) {
-            // Setup and install the trust manager
-            try {
-                final SSLContext sc = SSLContext.getInstance("SSL");
-                sc.init(null, new TrustManager[] { new FigoTrustManager() }, new java.security.SecureRandom());
-                ((HttpsURLConnection) connection).setSSLSocketFactory(sc.getSocketFactory());
-            } catch (NoSuchAlgorithmException e) {
-                throw new IOException("Connection setup failed", e);
-            } catch (KeyManagementException e) {
-                throw new IOException("Connection setup failed", e);
-            }
-        }
-
-        connection.setRequestMethod(method);
-        connection.setRequestProperty("Authorization", "Bearer " + this.access_token);
-        connection.setRequestProperty("Accept", "application/json");
-        connection.setRequestProperty("Content-Type", "application/json");
-
-        // add payload
-        if (data != null) {
-            String encodedData = createGson().toJson(data);
-
-            connection.setDoOutput(true);
-            connection.getOutputStream().write(encodedData.getBytes(UTF_8));
-        }
-
-        // process response
-        int code = connection.getResponseCode();
-        if (code >= 200 && code < 300) {
-            if (typeOfT == null)
-                return null;
-            else
-                return handleResponse(connection.getInputStream(), typeOfT);
-        } else if (code == 400) {
-            throw new FigoException((ErrorResponse) handleResponse(connection.getErrorStream(), FigoException.ErrorResponse.class));
-        } else if (code == 401) {
-            throw new FigoException("access_denied", "Access Denied");
-        } else if (code == 404) {
-            return null;
-        } else {
-            // return decode(connection.getErrorStream(), resultType);
-            throw new FigoException("internal_server_error", "We are very sorry, but something went wrong");
-        }
-    }
-
-    /**
-     * Instantiate the GSON class. Meant to be overridden in order to provide custom Gson settings.
-     * 
-     * @return GSON instance
-     */
-    protected Gson createGson() {
-        return GsonAdapter.createGson();
-    }
+//    /**
+//     * Instantiate the GSON class. Meant to be overridden in order to provide custom Gson settings.
+//     * 
+//     * @return GSON instance
+//     */
+//    protected Gson createGson() {
+//        return GsonAdapter.createGson();
+//    }
     
-    /**
-     * Handle the response of a request by decoding its JSON payload
-     * 
-     * @param stream
-     *            Stream containing the JSON data
-     * @param typeOfT
-     *            Type of the data to be expected
-     * @return Decoded data
-     */
-    private <T> T handleResponse(InputStream stream, Type typeOfT) {
-        // check whether decoding is actual requested
-        if (typeOfT == null)
-            return null;
-
-        // read stream body
-        Scanner s = new Scanner(stream, "UTF-8");
-        s.useDelimiter("\\A");
-        String body = s.hasNext() ? s.next() : "";
-        s.close();
-
-        // decode JSON payload
-        return createGson().fromJson(body, typeOfT);
-    }
+//    /**
+//     * Handle the response of a request by decoding its JSON payload
+//     * 
+//     * @param stream
+//     *            Stream containing the JSON data
+//     * @param typeOfT
+//     *            Type of the data to be expected
+//     * @return Decoded data
+//     */
+//    private <T> T handleResponse(InputStream stream, Type typeOfT) {
+//        // check whether decoding is actual requested
+//        if (typeOfT == null)
+//            return null;
+//
+//        // read stream body
+//        Scanner s = new Scanner(stream, "UTF-8");
+//        s.useDelimiter("\\A");
+//        String body = s.hasNext() ? s.next() : "";
+//        s.close();
+//
+//        // decode JSON payload
+//        return createGson().fromJson(body, typeOfT);
+//    }
 
     /**
      * Get the current figo Account
@@ -477,7 +450,7 @@ public class FigoSession {
      * @param count
      *            limit the number of returned transactions
      * @param offset
-     *            which offset into the result set should be used to determin the first transaction to return (useful in combination with count)
+     *            which offset into the result set should be used to determine the first transaction to return (useful in combination with count)
      * @param include_pending
      *            this flag indicates whether pending transactions should be included in the response; pending transactions are always included as a complete
      *            set, regardless of the `since` parameter
@@ -497,7 +470,7 @@ public class FigoSession {
      * @param count
      *            limit the number of returned transactions
      * @param offset
-     *            which offset into the result set should be used to determin the first transaction to return (useful in combination with count)
+     *            which offset into the result set should be used to determine the first transaction to return (useful in combination with count)
      * @param include_pending
      *            this flag indicates whether pending transactions should be included in the response; pending transactions are always included as a complete
      *            set, regardless of the `since` parameter
@@ -530,7 +503,7 @@ public class FigoSession {
      * Retrieve a specific transaction by ID
      * 
      * @param accountId
-     *            ID of the account on which the transaction occured
+     *            ID of the account on which the transaction occurred
      * @param transactionId
      *            the figo ID of the specific transaction
      * @return Transaction or null
@@ -925,11 +898,11 @@ public class FigoSession {
     public String submitPayment(Payment payment, String tanSchemeId, String state, String redirectUri) throws FigoException, IOException {
         TaskTokenResponse response = this.queryApi("/rest/accounts/" + payment.getAccountId() + "/payments/" + payment.getPaymentId() + "/submit",
                 new SubmitPaymentRequest(tanSchemeId, state, redirectUri), "POST", TaskTokenResponse.class);
-        return apiEndpoint + "/task/start?id=" + response.task_token;
+        return getApiEndpoint() + "/task/start?id=" + response.task_token;
     }
 
     /**
-     * URL to trigger a synchronisation. The user should open this URL in a web browser to synchronize his/her accounts with the respective bank servers. When
+     * URL to trigger a synchronization. The user should open this URL in a web browser to synchronize his/her accounts with the respective bank servers. When
      * the process is finished, the user is redirected to the provided URL.
      * 
      * @param state
@@ -941,7 +914,7 @@ public class FigoSession {
      */
     public String getSyncURL(String state, String redirect_url) throws FigoException, IOException {
         TaskTokenResponse response = this.queryApi("/rest/sync", new SyncTokenRequest(state, redirect_url), "POST", TaskTokenResponse.class);
-        return apiEndpoint + "/task/start?id=" + response.task_token;
+        return getApiEndpoint() + "/task/start?id=" + response.task_token;
     }
     
     /**
